refacto: déplacement de l'Exercice dans l'app entreprises
